Move tweet formatting into a separate helper class
