+++ app/helpers/social_helper.rb
@@ -1,0 +1,4 @@
+module SocialHelper
+  include SocialHelper::TwitterMethods
+  include SocialHelper::TumblrMethods
+end
